fix all register controller bugs

+++ src/utils/cloudinary.js
@@ -2,20 +2,26 @@ import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
 cloudinary.config({ 
-        cloud_name: process.env.CLOUSINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUSINARY_API_KEY, 
-        api_secret: process.env.CLOUSINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
+        cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+        api_key: process.env.CLOUDINARY_API_KEY, 
+        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if(!localFilePath) return null
-        const response = cloudinary.uploader.upload(localFilePath,{ resource_type: "auto" }) 
-        console.log("File uploaded on cloudinary", response.url)
+        const response = await cloudinary.uploader.upload(localFilePath,{ resource_type: "auto" }) 
+        //console.log("File uploaded on cloudinary", response.url)
+        fs.unlinkSync(localFilePath)
         return response
     } catch (error) {
-        fs.unlinkSync(localFilePath)
-        return null
+        console.error("Cloudinary upload failed:", error);
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
+        return null;
     }
 }
